requirements.txt
google-adk
vertexai
python-dotenv
streamlit
requests
fastapi
uvicorn
google-generativeai
